Remove casts from SyscallReturnCode
The updated version of vmm-sys-util allows us to use different types
with SyscallReturnCode, so many places don't need type conversions for
the return codes anymore.

+++ src/vmm/src/io_uring/queue/submission.rs
@@ -142,7 +142,7 @@ impl SubmissionQueue {
                 flags,
                 std::ptr::null::<libc::sigset_t>(),
             )
-        } as libc::c_int)
+        })
         .into_result()?;
         // It's safe to convert to u32 since the syscall didn't return an error.
         let submitted = u32::try_from(submitted).unwrap();
